% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obifasta.R
\name{extract_features}
\alias{extract_features}
\title{Extract Features from JSON Strings}
\usage{
extract_features(sequences, ..., verbose = is_robi_verbose())
}
\arguments{
\item{sequences}{A tibble with a column named "features" containing JSON strings.}

\item{...}{A character vector or several characters values specifying the keys to extract from the JSON strings.}

\item{verbose}{A logical indicating whether to show the progress bar.}
}
\value{
A tibble with the extracted features bound as additional columns.
}
\description{
This function extracts features from JSON strings within a tibble and binds the features to the tibble as additional columns.
}
\examples{
\dontrun{
sequences <- tibble(
features = c('{"a": 1, "b": 2}', '{"a": 3, "b": 4}')
)
extract_features(sequences, c("a", "b"))
}
}
