% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obifasta.R
\name{extract_readcount}
\alias{extract_readcount}
\title{Extract Read Counts}
\usage{
extract_readcount(sequences, key = "merged_sample")
}
\arguments{
\item{sequences}{A data frame containing the sequences information, as provided by the read_obifasta function.}

\item{key}{A character string specifying the key name in the JSON data.}
}
\value{
A sparse matrix representing the read counts for each sample and each MOTU.
}
\description{
This function extracts the read count data from the input sequences.
}
\examples{
file <- system.file("extdata", "sample.fasta", package="ROBIFastread")
seq <- read_obifasta(file)
m <- extract_readcount(seq)

}
