% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obifastq.R
\name{read_obifastq}
\alias{read_obifastq}
\title{Read an OBI-formatted FASTQ file}
\usage{
read_obifastq(file, keys = NULL, verbose = is_robi_verbose())
}
\arguments{
\item{file}{a character string specifying the path to the FASTQ file.}

\item{keys}{an optional character vector of keys to extract from the definition line.}

\item{verbose}{a logical flag indicating whether to print messages during the processing.}
}
\value{
a tibble with columns "id", "features", "definition", and "sequence".
If keys is provided, the additional columns for the specified keys will be added.
}
\description{
This function reads an OBI-formatted FASTQ file and returns a tibble containing
the sequence information. The FASTQ definition line is expected to contain
key-value pairs expressed in the JSON format. If the keys argument is provided,
the specified keys will be extracted from the JSON annotations and added as columns to the tibble.
}
\examples{
file <- system.file("extdata", "sample.fasta", package="ROBIFastread")
read_obifastq(file)

}
