% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robimessage.R
\name{robimessage}
\alias{robimessage}
\title{Writes a message on the console in verbose mode}
\usage{
robimessage(message, ..., .envir = parent.frame())
}
\arguments{
\item{message}{a string template sent to the \code{glue::glue} preprocessor.}

\item{...}{other values to be concatenated to the end of the message}

\item{.envir}{the environment passed to the  \code{glue::glue} preprocessor}
}
\description{
The message is written to the console if the code
is executed in the verbose mode. The verbose mode
can be specified by declaring a \code{verbose} logical
value to \code{TRUE}. If no \code{verbose} variable is defined,
the result of the \code{ROBITools2::is_robi_verbose}
function is used to decide of the verbose status.
}
\examples{
verbose <- TRUE
robimessage("Hello world !")
verbose <- FALSE
robimessage("Hello world !")
rm(verbose)
v_mode <- is_robi_verbose()
v_mode
robimessage("Hello world !")
set_robi_verbose(FALSE)
is_robi_verbose()
robimessage("Hello world !")
set_robi_verbose(TRUE)
is_robi_verbose()
robimessage("Hello world !")
set_robi_verbose(v_mode)
}
