% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robimessage.R
\name{set_robi_verbose}
\alias{set_robi_verbose}
\alias{is_robi_verbose}
\title{Sets the verbose status}
\usage{
set_robi_verbose(state)

is_robi_verbose(.envir = parent.frame())
}
\arguments{
\item{state}{a logical value indicating the new verbose mode}
}
\value{
a logical value
}
\description{
Mainly indicates if messages are sent to the user terminal
to indicate computation progression.
}
\examples{

state <- getOption("ROBITools2.verbose")
is_robi_verbose()
set_robi_verbose(TRUE)
is_robi_verbose()
set_robi_verbose(FALSE)
is_robi_verbose()
set_robi_verbose(state)

is_robi_verbose()

}
